+++ trunk/bacnet-stack/demo/object/ai.c
@@ -133,7 +133,7 @@ char *Analog_Input_Name(
     static char text_string[32] = "";   /* okay for single thread */
 
     if (object_instance < MAX_ANALOG_INPUTS) {
-        sprintf(text_string, "ANALOG INPUT %u", object_instance);
+        sprintf(text_string, "ANALOG INPUT %lu", (unsigned long)object_instance);
         return text_string;
     }
