+++ trunk/bacnet-stack/demo/object/bacfile.c
@@ -203,7 +203,7 @@ int bacfile_encode_property_apdu(
                 object_instance);
             break;
         case PROP_OBJECT_NAME:
-            sprintf(text_string, "FILE %d", object_instance);
+            sprintf(text_string, "FILE %lu", (unsigned long)object_instance);
             characterstring_init_ansi(&char_string, text_string);
             apdu_len =
                 encode_application_character_string(&apdu[0], &char_string);
